proper types for printing
git-svn-id: https://asn1c.svn.sourceforge.net/svnroot/asn1c/trunk@1371 59561ff5-6e30-0410-9f3c-9617f08c8826

+++ skeletons/per_opentype.c
@@ -183,7 +183,7 @@ uper_open_type_get_complex(asn_codec_ctx_t *ctx, asn_TYPE_descriptor_t *td,
 		case 0: break;
 		default:
 			ASN_DEBUG("Non-blank padding (%d bits 0x%02x)",
-				padding, pvalue);
+				padding, (int)pvalue);
 			UPDRESTOREPD;
 			_ASN_DECODE_FAILED;
 		}
